FAR-123/refactor: 관광지 리팩토링
