+++ src/main/java/com/ssafy/faraway/domain/attraction/controller/AttractionController.java
@@ -1,8 +1,8 @@
 package com.ssafy.faraway.domain.attraction.controller;
 
-import com.ssafy.faraway.domain.attraction.dto.req.AttractionSearchCondition;
-import com.ssafy.faraway.domain.attraction.dto.res.AttractionResponse;
-import com.ssafy.faraway.domain.attraction.dto.res.GugunResponse;
+import com.ssafy.faraway.domain.attraction.repository.dto.AttractionSearchCondition;
+import com.ssafy.faraway.domain.attraction.controller.dto.AttractionResponse;
+import com.ssafy.faraway.domain.attraction.controller.dto.GugunResponse;
 import com.ssafy.faraway.domain.attraction.entity.Sido;
 import com.ssafy.faraway.domain.attraction.service.AttractionQueryService;
 import com.ssafy.faraway.domain.attraction.service.AttractionService;
@@ -18,16 +18,15 @@ import java.util.List;
 @Slf4j
 @RequiredArgsConstructor
 @RequestMapping("/attraction")
+@CrossOrigin(originPatterns = "*")
 @Api(tags = "attraction")
 public class AttractionController {
     private final AttractionService attractionService;
     private final AttractionQueryService attractionQueryService;
     @GetMapping("/sido")
     public List<Sido> findAllSido() throws IOException {
-        List<Sido> list = null;
-        list = attractionService.findAllSido();
+        List<Sido> list = attractionService.findAllSido();
         if(list == null || list.size() == 0) {
-//                throw new CustomException()
             throw new IOException();
         }
         return list;
@@ -35,8 +34,7 @@ public class AttractionController {
 
     @GetMapping("/gugun/{sidoCode}")
     public List<GugunResponse> searchGugunBySidoCode(@PathVariable int sidoCode) throws IOException{
-        List<GugunResponse> list = null;
-        list = attractionQueryService.searchGugunBySidoCode(sidoCode);
+        List<GugunResponse> list = attractionQueryService.searchGugunBySidoCode(sidoCode);
         if(list == null || list.size() == 0) {
             throw new IOException();
         }
@@ -44,10 +42,9 @@ public class AttractionController {
 
     }
 
-    @GetMapping("")
+    @GetMapping
     public List<AttractionResponse> searchByCondition(@RequestBody AttractionSearchCondition condition) throws IOException{
-        List<AttractionResponse> list = null;
-        list = attractionQueryService.searchByCondition(condition);
+        List<AttractionResponse> list = attractionQueryService.searchByCondition(condition);
         if(list == null || list.size() == 0) {
             throw new IOException();
         }
