+++ src/main/java/com/ssafy/faraway/domain/attraction/controller/dto/GugunResponse.java
@@ -1,0 +1,16 @@
+package com.ssafy.faraway.domain.attraction.controller.dto;
+
+import lombok.Builder;
+import lombok.Data;
+
+@Data
+public class GugunResponse {
+    int gugunCode;
+    String gugunName;
+
+    @Builder
+    public GugunResponse(int gugunCode, String gugunName) {
+        this.gugunCode = gugunCode;
+        this.gugunName = gugunName;
+    }
+}
