+++ src/main/java/com/ssafy/faraway/domain/attraction/controller/dto/SidoResponse.java
@@ -1,0 +1,16 @@
+package com.ssafy.faraway.domain.attraction.controller.dto;
+
+import lombok.Builder;
+import lombok.Data;
+
+@Data
+public class SidoResponse {
+    private Integer sidoCode;
+    private String sidoName;
+
+    @Builder
+    public SidoResponse(Integer sidoCode, String sidoName) {
+        this.sidoCode = sidoCode;
+        this.sidoName = sidoName;
+    }
+}
