--- src/main/java/com/ssafy/faraway/domain/attraction/dto/res/AttractionResponse.java
@@ -1,15 +1,0 @@
-package com.ssafy.faraway.domain.attraction.dto.res;
-
-import lombok.Data;
-
-@Data
-public class AttractionResponse {
-    private int contentId;
-    private String title;
-    private String addr1;
-    private String zipcode;
-    private String tel;
-    private String firstImage;
-    private double latitude;
-    private double longitude;
-}
