--- src/main/java/com/ssafy/faraway/domain/attraction/dto/res/GugunResponse.java
@@ -1,9 +1,0 @@
-package com.ssafy.faraway.domain.attraction.dto.res;
-
-import lombok.Data;
-
-@Data
-public class GugunResponse {
-    int gugunCode;
-    String gugunName;
-}
