--- src/main/java/com/ssafy/faraway/domain/attraction/dto/res/SidoResponse.java
@@ -1,9 +1,0 @@
-package com.ssafy.faraway.domain.attraction.dto.res;
-
-import lombok.Data;
-
-@Data
-public class SidoResponse {
-    private Integer sidoCode;
-    private String sidoName;
-}
