+++ src/main/java/com/ssafy/faraway/domain/attraction/repository/dto/AttractionSearchCondition.java
@@ -1,4 +1,4 @@
-package com.ssafy.faraway.domain.attraction.dto.req;
+package com.ssafy.faraway.domain.attraction.repository.dto;
 
 import lombok.Data;
