+++ src/main/java/com/ssafy/faraway/domain/attraction/service/AttractionQueryService.java
@@ -1,8 +1,8 @@
 package com.ssafy.faraway.domain.attraction.service;
 
-import com.ssafy.faraway.domain.attraction.dto.req.AttractionSearchCondition;
-import com.ssafy.faraway.domain.attraction.dto.res.AttractionResponse;
-import com.ssafy.faraway.domain.attraction.dto.res.GugunResponse;
+import com.ssafy.faraway.domain.attraction.repository.dto.AttractionSearchCondition;
+import com.ssafy.faraway.domain.attraction.controller.dto.AttractionResponse;
+import com.ssafy.faraway.domain.attraction.controller.dto.GugunResponse;
 
 import java.util.List;
